Move validation components to its own folder
